requirements.txt
PyNaCl>=1.5.0
qrcode>=7.4.2
